refactor: separate class
