+++ src/main/kotlin/blackjack/BlackjackApplication.kt
@@ -1,0 +1,18 @@
+package blackjack
+
+import blackjack.controller.BlackjackController
+
+class BlackjackApplication(
+    private val controller: BlackjackController = BlackjackController(),
+) {
+    fun play() {
+        controller.run {
+            initGame()
+            dealCards()
+            playPlayerTurn()
+            playDealerTurn()
+            showParticipantsSummary()
+            setResult()
+        }
+    }
+}
